requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-filter==1.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
Markdown==2.6.11
mysqlclient==1.3.12
Pillow==5.0.0
Pygments==2.2.0
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
uWSGI==2.0.17
